Autenticando as rotas de memória com JWT

+++ src/routes/memories.ts
@@ -5,8 +5,15 @@ import { prisma } from '../lib/prisma'
 
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (req) => {
+    await req.jwtVerify()
+  })
+
+  app.get('/memories', async (req) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub
+      },
       orderBy: {
         createdAt: 'asc'
       }
@@ -19,7 +26,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     }))
   })
 
-  app.get('/memories/:id', async (req) => {
+  app.get('/memories/:id', async (req, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -29,6 +36,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id }
     })
+
+    if(!memory.isPublic && memory.userId !== req.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -47,14 +58,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'd081bf08-428d-44d3-b182-04d61a65b049'
+        userId: req.user.sub
       }
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (req) => {
+  app.put('/memories/:id', async (req, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -69,7 +80,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(req.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id, 
+      }
+    })
+    
+    if(memory.userId !== req.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
@@ -81,12 +102,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (req) => {
+  app.delete('/memories/:id', async (req, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(req.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id, 
+      }
+    })
+    
+    if(memory.userId !== req.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: { id }
